Move key utils into auth module
